Reorganized Graph Class
